Add intermediate caching for EMD ref with constant color
